reorganized loaders into a subpackage
